requirements.txt
flask
ultralytics
pillow==9.5.0
easyocr
Gunicorn
